Prepared for pull request
