+++ index.html
@@ -7,12 +7,14 @@
 
   <script src="lib/D3/d3.min.js"></script>
 
+  <script src="js/classes/shapes/Circle.js"></script>
+  <script src="js/classes/shapes/Point.js"></script>
+  <script src="js/classes/shapes/Rect.js"></script>
+
   <!-- Buttons -->
   <script src="js/classes/gui/buttons/Button.js"></script>
   <script src="js/classes/gui/buttons/BorderedButton.js"></script>
-
-
-  <script src="js/classes/gui/ItemButton.js"></script>
+  <script src="js/classes/gui/buttons/ItemButton.js"></script>
 
   <script src="js/classes/gui/Inventory.js"></script>
   <script src="js/classes/gui/Slot.js"></script>
@@ -20,14 +22,11 @@
   <script src="js/classes/gui/Store.js" charset="utf-8"></script>
   <script src="js/classes/gui/ToolTip.js" charset="utf-8"></script>
   <script src="js/classes/gui/ValueBox.js"></script>
-  <script src="js/classes/shapes/Circle.js"></script>
-  <script src="js/classes/shapes/Point.js"></script>
-  <script src="js/classes/shapes/Rect.js"></script>
+
   <script src="js/classes/world_objects/Drop.js"></script>
   <script src="js/classes/world_objects/GameObject.js"></script>
   <script src="js/classes/world_objects/Snappable.js"></script>
   <script src="js/classes/world_objects/Pipe.js"></script>
-  <script src="js/classes/world_objects/Pipe2.js"></script>
   <script src="js/classes/world_objects/Pump.js"></script>
   <script src="js/classes/world_objects/Valve.js"></script>
   <script src="js/classes/world_objects/Faucet.js"></script>
@@ -71,7 +70,6 @@
 
 		$(document).ready(function() {
       game = new Game();
-
 		});
    </script>
  </head>
