+++ js/classes/gui/buttons/BorderedButton.js
@@ -1,35 +1,30 @@
 /*
+  BorderedButton
   A button with a border around it
-
 */
+class BorderedButton extends Button {
+  constructor(position, width, height, borderWidth) {
+    super(position, width, height)
 
-function BorderedButton(position, width, height, borderWidth) {
+    this.borderWidth = borderWidth;
 
-  Button.call(this, position, width, height);
-
-  this.borderWidth = borderWidth;
-
-  var mainSVG = d3.select("body").select("svg")
-	this.svg = {
-    border: mainSVG.append("rect"),
-		innerRect: mainSVG.append("rect"),
-		label: mainSVG.append("text"),
-		clickBox: mainSVG.append("rect")
-	};
-
-}
-
-BorderedButton.prototype = Object.create(Button.prototype)
-BorderedButton.prototype.constructor = BorderedButton
+    var mainSVG = d3.select("body").select("svg")
+  	this.svg = {
+      border: mainSVG.append("rect"),
+  		innerRect: mainSVG.append("rect"),
+  		label: mainSVG.append("text"),
+  		clickBox: mainSVG.append("rect")
+  	};
+  }
 
 
-/*
-	Creates the graphics for the background of
-	the button
-*/
-BorderedButton.prototype.createBackgroundSVG = function () {
-
-		// border
+  /**
+    createBackgroundSVG()
+  	@description Creates the graphics for the background of
+  	the button
+  */
+  createBackgroundSVG() {
+    // border
 		this.svg.border.attr("x", this.position.x);
 		this.svg.border.attr("y", this.position.y);
 		this.svg.border.attr("width", this.width);
@@ -42,5 +37,7 @@ BorderedButton.prototype.createBackgroundSVG = function () {
 		this.svg.innerRect.attr("width", this.width - this.borderWidth * 2);
 		this.svg.innerRect.attr("height", this.height - this.borderWidth * 2);
 		this.svg.innerRect.style("fill", "white")
+  }
+}
 
-};
+//export default BorderedButton
