--- js/classes/world_objects/Snappable2.js
@@ -1,11 +1,0 @@
-
-
-class Snappable2 extends GameObject {
-  constructor() {
-    this.center = center
-    this.orientation = "horizontal"
-
-  }
-
-
-}
