--- js2/classes/gui/Store.js
@@ -1,13 +1,0 @@
-/*
-  A place where a player can purchase, and sell his or her
-  items.
-
-*/
-class Store {
-  constructor() {
-    this.forSaleItems = []; // the items available that the player can
-                            // buy
-    this.sellableItems = []; // the players own items that he or she can sell
-
-  }
-}
