--- js2/classes/shapes/Point.js
@@ -1,3 +1,0 @@
-function Distance(point1, point2) {
-  return Math.sqrt(Math.pow(point1.x - point2.x, 2) + Math.pow(point1.y - point2.y, 2))
-}
