--- js2/classes/world_objects/tanks/SellTank.js
@@ -1,1 +1,0 @@
-
