--- js2/classes/world_objects/tanks/SidedTank.js
@@ -1,4 +1,0 @@
-/*
-  SidedTank: a sided tank is a tank with the added ability
-  of choosing which side or sides are open.
-*/
