--- js2/loader.js
@@ -1,3 +1,0 @@
-/*
-  Loader.js loads all the script files into the html document.
-*/
