--- testPage2.html
@@ -1,76 +1,0 @@
-<!DOCTYPE html>
-<html>
- <head>
-  <title>
-   Chemistry
-  </title>
-
-  <script src="lib/D3/d3.min.js"></script>
-
-  <!-- Buttons -->
-  <script src="js2/classes/gui/buttons/Button.js"></script>
-  <script src="js2/classes/gui/buttons/BorderedButton.js"></script>
-  <script src="js2/classes/gui/buttons/ItemButton.js"></script>
-
-  <script src="js2/classes/gui/Inventory.js"></script>
-
-
-
-  <!--
-
-		TODO: Pipes to connect the two tanks
-		TODO: Animation of liquid flowing through the pipes
-		TODO: Pump (whenever a pump is clicked it produces a square drop of liquid which moves down from the opening
-					of the pump until it encounters a tank. At that point, it enters the tank, and fills that tank 1 unit.)
-		TODO: Multiple tanks
-		TODO: Pump to pump liquids out of tanks
-		TODO: Prevent text from being selected.
-		TODO: Prevent pipes from overlapping.
-		TODO: Liquid not entering tank.
-	-->
-  <!--
-		Main Idea:
-			A game about creating the most efficient fuel
-
-
-			You have a pump to collect liquid. Every click you get one drop of liquid.
-			Later, you can use your liquid to power an engine that will automatically pump liquid
-	-->
-  <link charset="utf-8" href="css/style.css" media="screen" rel="stylesheet" title="no title"></link>
-   <script src="lib/jquery-1.11.3.min.js"></script>
-   <script charset="utf-8" src="lib/util.js"></script>
-   <!--
-		Tests
-			-Rect.contains
-			-Rect.intersects
-	-->
-   <script type="text/javascript">
-    var game = null;
-
-		$(document).ready(function() {
-      var inventory = new Inventory()
-
-
-      var button = new ItemButton({
-          x: 300,
-          y: 300
-        },
-        100, // width
-        50, // height
-        2 // borderWidth
-      );
-
-    button.setText("Test")
-    button.setOnClick(function() {
-      console.log("Clicked");
-    })
-    button.createSVG();
-
-		});
-   </script>
- </head>
- <body>
-  <svg id="main" height="600" width="1750">
-  </svg>
- </body>
-</html>
